[refactor] 모듈 분리
