+++ app/build.gradle.kts
@@ -44,6 +44,7 @@ android {
 
 dependencies {
     implementation(project(":domain"))
+    implementation(project(":pingudi"))
 
     implementation("androidx.core:core-ktx:1.10.1")
     implementation("androidx.appcompat:appcompat:1.6.1")
@@ -69,7 +70,6 @@ dependencies {
     // Robolectric
     testImplementation("org.robolectric:robolectric:4.9")
     // Reflection
-    implementation("javax.inject", "javax.inject", "1")
     implementation("org.jetbrains.kotlin", "kotlin-reflect", "1.8.21")
 
     // mockk
