+++ app/src/main/java/woowacourse/shopping/data/InDiskCartRepository.kt
@@ -1,9 +1,9 @@
 package woowacourse.shopping.data
 
+import com.example.pingudi.annotation.InDisk
 import woowacourse.shopping.data.mapper.toDomain
 import woowacourse.shopping.data.mapper.toEntity
 import woowacourse.shopping.data.repository.CartRepository
-import woowacourse.shopping.di.InDisk
 import woowacourse.shopping.model.CartProduct
 import woowacourse.shopping.model.Product
