--- app/src/main/java/woowacourse/shopping/di/InMemory.kt
@@ -1,3 +1,0 @@
-package woowacourse.shopping.di
-
-annotation class InMemory
