+++ app/src/main/java/woowacourse/shopping/di/ViewModelLazy.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelLazy
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.ViewModelStoreOwner
+import com.example.pingudi.Injector
 import woowacourse.shopping.ShoppingApplication
 
 @MainThread
