+++ app/src/test/java/woowacourse/shopping/di/InjectorTest.kt
@@ -14,8 +14,8 @@ class InjectorTest {
     @Test
     fun `주어진 클래스 타입에 맞게 인스턴스를 생성해서 반환한다`() {
         // when
-        Container.addInstance(Fake.DefaultFakeRepository::class, null)
-        val injector = Injector(Container)
+        com.example.pingudi.Container.addInstance(Fake.DefaultFakeRepository::class, null)
+        val injector = com.example.pingudi.Injector(com.example.pingudi.Container)
         val actual = injector.createInstance(Fake.FakeViewModel::class)
 
         // then
@@ -24,15 +24,15 @@ class InjectorTest {
 
     @Test(expected = IllegalStateException::class)
     fun `주어진 클래스 타입의 주생성자를 가져올 수 없다면 오류가 발생한다`() {
-        val injector = Injector(Container)
+        val injector = com.example.pingudi.Injector(com.example.pingudi.Container)
         injector.createInstance(FakeViewModel2::class)
     }
 
     @Test(expected = IllegalArgumentException::class)
     fun `container에 없는 타입을 요구할 경우 오류가 발생한다`() {
         // given
-        Container.addInstance(Fake.DefaultFakeRepository::class, null)
-        val injector = Injector(Container)
+        com.example.pingudi.Container.addInstance(Fake.DefaultFakeRepository::class, null)
+        val injector = com.example.pingudi.Injector(com.example.pingudi.Container)
 
         // when
         injector.createInstance(FakeViewModel3::class)
