+++ pingudi/build.gradle.kts
@@ -1,0 +1,23 @@
+plugins {
+    id("java-library")
+    id("org.jetbrains.kotlin.jvm")
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_11
+    targetCompatibility = JavaVersion.VERSION_11
+}
+
+dependencies {
+    // Reflection
+    implementation("org.jetbrains.kotlin", "kotlin-reflect", "1.8.21")
+
+    testImplementation("org.junit.jupiter", "junit-jupiter", "5.8.2")
+    testImplementation("org.assertj", "assertj-core", "3.22.0")
+}
+
+tasks {
+    test {
+        useJUnitPlatform()
+    }
+}
