+++ pingudi/src/main/java/com/example/pingudi/annotation/InDisk.kt
@@ -1,0 +1,3 @@
+package com.example.pingudi.annotation
+
+annotation class InDisk
