+++ pingudi/src/main/java/com/example/pingudi/annotation/InMemory.kt
@@ -1,0 +1,3 @@
+package com.example.pingudi.annotation
+
+annotation class InMemory
