+++ pingudi/src/main/java/com/example/pingudi/annotation/InjectField.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.di
+package com.example.pingudi.annotation
 
 @Target(AnnotationTarget.PROPERTY)
 annotation class InjectField
